# src/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  model.ApiError:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  model.CreateConversationRequest:
    properties:
      episode_id:
        type: string
    required:
    - episode_id
    type: object
  model.CreateConversationResponse:
    properties:
      conversation_id:
        type: string
    type: object
  model.EpisodeCard:
    properties:
      cover_url:
        type: string
      duration_seconds:
        type: integer
      id:
        type: string
      podcast_name:
        type: string
      published_at:
        type: string
      title:
        type: string
    type: object
  model.EpisodeDetail:
    properties:
      cover_url:
        type: string
      duration_seconds:
        type: integer
      id:
        type: string
      podcast_name:
        type: string
      published_at:
        type: string
      title:
        type: string
    type: object
  model.EpisodeDetailResponse:
    properties:
      episode:
        $ref: '#/definitions/model.EpisodeDetail'
    type: object
  model.EpisodeListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/model.EpisodeCard'
        type: array
      next_cursor:
        type: string
      total:
        type: integer
    type: object
  model.FactCheckClaim:
    properties:
      claim:
        type: string
      explanation:
        type: string
      id:
        type: string
      sources:
        items:
          type: string
        type: array
      start_time:
        type: integer
      verdict:
        type: string
    type: object
  model.FactChecksResponse:
    properties:
      claims:
        items:
          $ref: '#/definitions/model.FactCheckClaim'
        type: array
      episode_id:
        type: string
    type: object
  model.HealthStatus:
    properties:
      database:
        type: string
      minio:
        type: string
      status:
        type: string
    type: object
  model.SendMessageContext:
    properties:
      current_time:
        type: integer
    type: object
  model.SendMessageRequest:
    properties:
      context:
        $ref: '#/definitions/model.SendMessageContext'
      text:
        maxLength: 10000
        type: string
    required:
    - text
    type: object
  model.TopicCard:
    properties:
      emotion:
        type: string
      id:
        type: string
      start_time:
        type: integer
      summary:
        type: string
      topic:
        type: string
    type: object
  model.TopicsResponse:
    properties:
      episode_id:
        type: string
      topics:
        items:
          $ref: '#/definitions/model.TopicCard'
        type: array
    type: object
  model.TranscriptLine:
    properties:
      has_fact_flag:
        type: boolean
      id:
        type: string
      start_time:
        type: integer
      text:
        type: string
    type: object
  model.TranscriptResponse:
    properties:
      episode_id:
        type: string
      lines:
        items:
          $ref: '#/definitions/model.TranscriptLine'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: Podcast analysis backend — episodes, topics, transcripts, fact-checks,
    chat.
  title: Audiolens API
  version: "1.0"
paths:
  /chat/conversations:
    post:
      consumes:
      - application/json
      description: Create a new chat conversation for an episode.
      parameters:
      - description: Episode ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateConversationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.CreateConversationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Neue Chat-Session starten
      tags:
      - chat
  /chat/conversations/{id}/messages:
    post:
      consumes:
      - application/json
      description: Send a message and receive a stubbed NDJSON streaming response.
      parameters:
      - description: Conversation ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: User message
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.SendMessageRequest'
      produces:
      - application/x-ndjson
      responses:
        "200":
          description: NDJSON stream of ChatStreamChunk
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Nachricht senden — Antwort kommt als Stream
      tags:
      - chat
  /episodes:
    get:
      description: Cursor-based paginated episode list with optional free-text search.
      parameters:
      - description: Free-text search on title and podcast name
        in: query
        name: q
        type: string
      - description: Cursor from previous response
        in: query
        name: cursor
        type: string
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.EpisodeListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Episode-Liste
      tags:
      - episodes
  /episodes/{id}:
    get:
      description: Get episode detail for the header area above the tabs.
      parameters:
      - description: Episode ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.EpisodeDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Episode-Detail
      tags:
      - episodes
  /episodes/{id}/fact-checks:
    get:
      description: Get fact-check claims for an episode.
      parameters:
      - description: Episode ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.FactChecksResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Fact-Check-Flags für die Sidebar
      tags:
      - fact-checks
  /episodes/{id}/sync:
    get:
      description: Server-Sent Events stream for playback synchronization (stubbed).
      parameters:
      - description: Episode ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Playback-Position per SSE
      tags:
      - sync
  /episodes/{id}/topics:
    get:
      description: Get topics for an episode. Returns 202 if analysis is not ready
        yet.
      parameters:
      - description: Episode ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TopicsResponse'
        "202":
          description: Analysis not ready
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Themen für den Themen-Tab
      tags:
      - topics
  /episodes/{id}/transcript:
    get:
      description: Get transcript lines for an episode. Lines are annotated with has_fact_flag.
      parameters:
      - description: Episode ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TranscriptResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Transkript-Zeilen
      tags:
      - transcript
  /health:
    get:
      description: Get the health status of the API, including database and MinIO
        connectivity.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HealthStatus'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/model.HealthStatus'
      summary: Health check
      tags:
      - health
swagger: "2.0"